Refactoring of CSDiff module and preparation for new implementation of consecutive lines textual merge strategy
