+++ src/main/java/br/ufpe/cin/mergers/util/CSDiffScript.java
@@ -1,0 +1,16 @@
+package br.ufpe.cin.mergers.util;
+
+public enum CSDiffScript {
+    ConsecutiveLines("/consecutive-lines.sh"),
+    CSDiff("/csdiff.sh");
+
+    private String path;
+
+    CSDiffScript(String path) {
+        this.path = path;
+    }
+
+    public String getPath() {
+        return this.path;
+    }
+}
